show

+++ frontend/src/components/PatentView.vue
@@ -207,6 +207,10 @@ export default {
       }
       this.commentModel.rate=this.starRate.toString();
       console.log(this.commentModel);
+      if (this.$route.params.userID == null) {
+        this.$Message.info("请先登录");
+        return;
+      }
       axios.post("/comment/", this.commentModel).then(res => {
         if (res.status == 201) {
           this.$Message.info("评论成功！");
@@ -220,26 +224,23 @@ export default {
       this.open = !this.open;
     },
     collectpatent() {
+      if (this.$route.params.userID == null) {
+        this.$Message.info("请先登录");
+        return;
+      }
       this.iscollect = !this.iscollect;
       if (this.iscollect) {
         axios.post("/collections/", {
-          params: {
-            userID: this.$route.params.userID
-          },
-          data: {
+         
             userID: this.$route.params.userID,
             resourceID: this.resourceID
-          }
         });
       } else {
         axios.delete("/collections/", {
-          params: {
-            userID: this.$route.params.userID
-          },
-          data: {
+          
             userID: this.$route.params.userID,
             resourceID: this.resourceID
-          }
+
         });
       }
     }
